fix: published deducer depends on vitest
