+++ .changeset/soft-walls-design.md
@@ -1,0 +1,5 @@
+---
+"@plutolang/static-deducer": patch
+---
+
+Fix: static deducer of the published version depends on vitest.
